requirements.txt
fastapi
uvicorn
gunicorn
google-generativeai
gTTS
pydantic
speechrecognition
python-dotenv